Change coordinates to double

+++ API/src/main/java/org/atelier2/entity/Point.java
@@ -11,8 +11,8 @@ package org.atelier2.entity;
  */
 public class Point {
     private long id;
-    private float lat;
-    private float lng;
+    private double lat;
+    private double lng;
     private String photo;
     private String description;
     private int difficulte;
@@ -25,19 +25,19 @@ public class Point {
         this.id = id;
     }
 
-    public float getLat() {
+    public double getLat() {
         return lat;
     }
 
-    public void setLat(float lat) {
+    public void setLat(double lat) {
         this.lat = lat;
     }
 
-    public float getLng() {
+    public double getLng() {
         return lng;
     }
 
-    public void setLng(float lng) {
+    public void setLng(double lng) {
         this.lng = lng;
     }
